backend/requirements.txt
fastapi>=0.136.0
uvicorn[standard]>=0.46.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
